third_party/blink/web_tests/fast/forms/label/label-contains-other-interactive-content-expected.txt

button
PASS document.activeElement.id is not "control"
details
PASS document.activeElement.id is not "control"
embed
PASS document.activeElement.id is not "control"
iframe
PASS document.activeElement.id is not "control"
label
PASS document.activeElement.id is not "control"
select
PASS document.activeElement.id is not "control"
textarea
PASS document.activeElement.id is not "control"


a
PASS document.activeElement.id is "control"
a[href]
PASS document.activeElement.id is not "control"
audio[controls]
PASS document.activeElement.id is not "control"
video
PASS document.activeElement.id is "control"
video[controls]
PASS document.activeElement.id is not "control"
img
PASS document.activeElement.id is "control"
img[usemap]
PASS document.activeElement.id is not "control"
object
FAIL document.activeElement.id should be control. Was object.
object[usemap]
PASS document.activeElement.id is not "control"
input
PASS document.activeElement.id is not "control"
PASS successfullyParsed is true

TEST COMPLETE
